+++ src/api/leadSources.js
@@ -4,8 +4,12 @@ const END_POINT = '/lead-sources';
 
 const getLeadSources = () => httpClient.get(END_POINT);
 const addLeadSources = (dataForm) => httpClient.post( END_POINT, dataForm);
+const updateLeadSources = (dataForm, id) => httpClient.put( END_POINT + '/' + id, dataForm);
+const getLeadSourceById = (id) => httpClient.get( END_POINT + '/' + id);
 
 export {
   getLeadSources,
-  addLeadSources
+  addLeadSources,
+  updateLeadSources,
+  getLeadSourceById
 }