+++ src/views/wny/leads/LeadSources.vue
@@ -13,8 +13,8 @@
 
         <v-client-table :columns="columns" :data="data" :options="options" :theme="theme" id="dataTable">
           <p slot="actions" slot-scope="props">
-            <a :href="'#/roles/' + props.row.id" class="icon-eye action-icon"></a>
-            <a :href="'#/roles/' + props.row.id + '/edit'" class="icon-pencil action-icon"></a>
+<!--            <a :href="'#/roles/' + props.row.id" class="icon-eye action-icon"></a>-->
+            <a :href="'#/admin/lead-sources/' + props.row.id + '/edit'" class="icon-pencil action-icon"></a>
             <a class="icon-trash" v-on:click="deleteRole(props.row.id)" style="cursor: pointer"></a>
           </p>